refactor: refactor from scratch
