+++ src_old/interface/speaker.py
