+++ src_old/setting/flask-src/static/js/setting.js
